Bugfix
Fixed a bug where adding a Steam game to the list of exceptions via
double middle mouse clicking would continue erroneously to iterate
through the table of games. This resulted in unnecessary work and in
attempts to access non-existing game objects. I also removed an
unnecessary for-loop that had been left in.