Added an extra Python script
Added a script that can be used to generate entries for videos if one
wishes to use Lauhdutin for their video library. Some modifications to
Lauhdutin are recommended such as stopping Lauhdutin from updating how
many hours an entry has been played for since that property is used by
the script for the duration of the video for sorting purposes. The
script creates entries that are instances of the new Custom platform.